# Conflicts:
#	data/src/main/graphql/com/multimoney/multimoney/graphql/accountsmart/mutation/GlobalRequest.graphql
#	data/src/main/graphql/com/multimoney/multimoney/graphql/security/query/CompanyByIdentification.graphql
#	data/src/main/java/com/multimoney/data/networking/SecurityApi.kt
#	data/src/main/java/com/multimoney/data/networking/SmartAccountApi.kt
#	data/src/main/java/com/multimoney/data/repository/SmartAccountRepositoryImpl.kt

+++ domain/src/main/java/com/multimoney/domain/interaction/accountsmart/MutationGlobalRequestUseCase.kt
@@ -19,7 +19,7 @@ interface MutationGlobalRequestUseCase {
         idAddressLevel2: Long,
         idAddressLevel3: Long,
         idEconomicActivity: Long,
-        income: Int,
+        income: Double,
         addressDetail: String,
         isPEP: Boolean,
         user: String,
